Allow for multiple now and next events
This allows for multiple now and next events in the response types used
in the client and adapter.

It does not allow the client to return multiple now/next events or allow
the CLI to display them, this will be done later.

Multiple events should not really be a common issue, but it did happen
in 2022, so might happen again, who knows.
May as well handle it correctly if it does happen though.

Re #13